+++ lib/src/classes/accessibility/EzImage.dart
