+++ lib/src/classes/accessibility/EzLink.dart
