+++ lib/src/classes/accessibility/EzText.dart
