+++ lib/src/classes/accessibility/EzTextStyle.dart
